+++ encoding/protobuf/stream_test.go
@@ -18,77 +18,206 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-package protobuf
+package protobuf_test
 
 import (
 	"context"
 	"errors"
-	"io/ioutil"
+	"io"
 	"testing"
 
-	"github.com/gogo/protobuf/proto"
-	"github.com/golang/mock/gomock"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
+	"go.uber.org/yarpc"
 	"go.uber.org/yarpc/api/transport"
-	"go.uber.org/yarpc/api/transport/transporttest"
-	"go.uber.org/yarpc/yarpcerrors"
+	"go.uber.org/yarpc/encoding/protobuf"
+	"go.uber.org/yarpc/encoding/protobuf/internal/testpb"
+	"go.uber.org/yarpc/internal/clientconfig"
+	"go.uber.org/yarpc/peer"
+	"go.uber.org/yarpc/peer/hostport"
+	"go.uber.org/yarpc/yarpctest"
 )
 
-func TestReadFromStreamDecodeError(t *testing.T) {
-	mockCtrl := gomock.NewController(t)
-	defer mockCtrl.Finish()
-
-	ctx := context.Background()
-	wantErr := errors.New("error")
-
-	stream := transporttest.NewMockStreamCloser(mockCtrl)
-	stream.EXPECT().ReceiveMessage(ctx).Return(&transport.StreamMessage{
-		Body: ioutil.NopCloser(readErr{err: wantErr}),
-	}, nil)
-	stream.EXPECT().Request().Return(
-		&transport.StreamRequest{
-			Meta: &transport.RequestMeta{
-				Encoding: Encoding,
-			},
+var protocolOptionsTable = []struct {
+	msg  string
+	opts []protobuf.ClientOption
+}{
+	{
+		msg: "protobuf",
+	},
+	{
+		msg: "json",
+		opts: []protobuf.ClientOption{
+			protobuf.UseJSON,
 		},
-	)
-
-	clientStream, err := transport.NewClientStream(stream)
-	require.NoError(t, err)
-
-	_, err = readFromStream(ctx, clientStream, func() proto.Message { return nil }, newCodec(nil /*AnyResolver*/))
-
-	assert.Equal(t, wantErr, err)
+	},
 }
 
-type readErr struct {
-	err error
+func TestUnary(t *testing.T) {
+	for _, tt := range protocolOptionsTable {
+		t.Run(tt.msg, func(t *testing.T) {
+			server := &testServer{}
+			procedures := testpb.BuildTestYARPCProcedures(server)
+			router := yarpc.NewMapRouter("test")
+			router.Register(procedures)
+
+			trans := yarpctest.NewFakeTransport()
+			pc := peer.NewSingle(hostport.Identify("1"), trans)
+			ob := trans.NewOutbound(pc, yarpctest.OutboundRouter(router))
+			cc := clientconfig.MultiOutbound("test", "test", transport.Outbounds{
+				Unary: ob,
+			})
+			client := testpb.NewTestYARPCClient(cc, tt.opts...)
+
+			ctx, cancel := context.WithCancel(context.Background())
+			defer cancel()
+
+			sent := &testpb.TestMessage{Value: "echo"}
+			received, err := client.Unary(ctx, sent)
+			require.NoError(t, err)
+			assert.Equal(t, sent, received)
+		})
+	}
 }
 
-func (r readErr) Read(p []byte) (n int, err error) {
-	return 0, r.err
+var _ testpb.TestYARPCServer = (*testServer)(nil)
+
+// testServer provides unary and streaming echo method implementations.
+type testServer struct{}
+
+func (s *testServer) Unary(ctx context.Context, msg *testpb.TestMessage) (*testpb.TestMessage, error) {
+	return msg, nil
 }
 
-func TestWriteToStreamInvalidEncoding(t *testing.T) {
-	mockCtrl := gomock.NewController(t)
-	defer mockCtrl.Finish()
+func (s *testServer) Duplex(str testpb.TestServiceDuplexYARPCServer) error {
+	for {
+		msg, err := str.Recv()
+		if err != nil {
+			return err
+		}
+		if msg.Value == "please explode" {
+			return errors.New("explosion occurred")
+		}
+		err = str.Send(msg)
+		if err != nil {
+			return err
+		}
+	}
+}
 
-	ctx := context.Background()
+func TestDuplexStream(t *testing.T) {
+	for _, tt := range protocolOptionsTable {
+		t.Run(tt.msg, func(t *testing.T) {
+			server := &testServer{}
+			procedures := testpb.BuildTestYARPCProcedures(server)
+			router := yarpc.NewMapRouter("test")
+			router.Register(procedures)
 
-	stream := transporttest.NewMockStreamCloser(mockCtrl)
-	stream.EXPECT().Request().Return(
-		&transport.StreamRequest{
-			Meta: &transport.RequestMeta{
-				Encoding: transport.Encoding("raw"),
-			},
+			trans := yarpctest.NewFakeTransport()
+			pc := peer.NewSingle(hostport.Identify("1"), trans)
+			ob := trans.NewOutbound(pc, yarpctest.OutboundRouter(router))
+			cc := clientconfig.MultiOutbound("test", "test", transport.Outbounds{
+				Stream: ob,
+			})
+			client := testpb.NewTestYARPCClient(cc, tt.opts...)
+
+			ctx, cancel := context.WithCancel(context.Background())
+			defer cancel()
+
+			str, err := client.Duplex(ctx)
+			require.NoError(t, err)
+
+			// Send a message.
+			sent := &testpb.TestMessage{Value: "echo"}
+			{
+				err := str.Send(sent)
+				require.NoError(t, err)
+			}
+
+			// Receive the echoed message.
+			{
+				msg, err := str.Recv()
+				require.NoError(t, err)
+				assert.Equal(t, sent, msg)
+			}
+
+			// Close the client side of the stream.
+			str.CloseSend()
+
+			// Verify that the server closes as well.
+			{
+				_, err := str.Recv()
+				require.Equal(t, err, io.EOF)
+			}
+		})
+	}
+}
+
+func TestStreamServerError(t *testing.T) {
+	table := []struct {
+		msg          string
+		clientCloses bool
+	}{
+		{
+			msg:          "client closes after error",
+			clientCloses: true,
 		},
-	)
+		{
+			msg:          "client returns immediately after error",
+			clientCloses: false,
+		},
+	}
 
-	clientStream, err := transport.NewClientStream(stream)
-	require.NoError(t, err)
+	for _, ct := range table {
+		t.Run(ct.msg, func(t *testing.T) {
+			for _, tt := range protocolOptionsTable {
+				t.Run(tt.msg, func(t *testing.T) {
+					server := &testServer{}
+					procedures := testpb.BuildTestYARPCProcedures(server)
+					router := yarpc.NewMapRouter("test")
+					router.Register(procedures)
 
-	err = writeToStream(ctx, clientStream, nil, newCodec(nil /*AnyResolver*/))
+					trans := yarpctest.NewFakeTransport()
+					pc := peer.NewSingle(hostport.Identify("1"), trans)
+					ob := trans.NewOutbound(pc, yarpctest.OutboundRouter(router))
+					cc := clientconfig.MultiOutbound("test", "test", transport.Outbounds{
+						Stream: ob,
+					})
+					client := testpb.NewTestYARPCClient(cc, tt.opts...)
 
-	assert.Equal(t, yarpcerrors.Newf(yarpcerrors.CodeInternal, "encoding.Expect should have handled encoding \"raw\" but did not"), err)
+					ctx, cancel := context.WithCancel(context.Background())
+					defer cancel()
+
+					str, err := client.Duplex(ctx)
+					require.NoError(t, err)
+
+					// Send a message.
+					sent := &testpb.TestMessage{Value: "please explode"}
+					{
+						err := str.Send(sent)
+						require.NoError(t, err)
+					}
+
+					// Receive the handler error on next receive.
+					{
+						msg, err := str.Recv()
+						require.Error(t, err)
+						assert.Nil(t, msg)
+					}
+
+					// Close the client side of the stream.
+					if ct.clientCloses {
+						str.CloseSend()
+					}
+
+					// Verify that the server closes as well.
+					{
+						msg, err := str.Recv()
+						require.Equal(t, err, io.EOF)
+						assert.Nil(t, msg)
+					}
+				})
+			}
+		})
+	}
 }